scripts/requirements.txt
# Scripts 가상환경 의존성
# 스크립트 실행에 필요한 패키지들

# 환경 변수 관리
python-dotenv>=1.0.0

# 데이터베이스
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP 요청
requests>=2.31.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# LangChain and LangGraph dependencies (for lawfirm_langgraph)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=1.0.0,<2.0.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
langchain-text-splitters>=1.0.0,<2.0.0

# Optional dependencies (for advanced features)
langsmith>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Korean NLP dependencies
konlpy>=0.6.0

# 유틸리티
tqdm>=4.66.0
colorama>=0.4.6,<0.5.0

# Pydantic (for data validation)
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0

# 로깅 및 모니터링
mlflow>=2.0.0

# 기타 유틸리티
pathlib2>=2.3.7; python_version < "3.4"

